mem/pyproject.toml
[project]
name = "mem"
version = "0.1.0"
description = "Video work tracking system with local AI"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Spencer Delcore", email = ""},
]

dependencies = [
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0,<2",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0.1",
    "duckdb>=0.9.0",
    "pytz>=2024.1",
    "tomli>=2.0.0",
    "imagehash>=4.3.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "PyGObject>=3.42.0,<3.50",
    "pycairo>=1.20.0,<1.28",
    "httpx>=0.28.1",
    "slowapi>=0.1.9",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "N", "SIM"]
ignore = ["E501", "B008", "B904"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-ra --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]
asyncio_mode = "auto"
